Leave values in strings if the ints are never used
Several status values are ultimately only used for display on
the web page, so don't bother converting back and forth to int;
that only complicates the code with intermediate values and
needless conversions.

Making this all consistent should make it simpler to drop
in a sending/parsing helper in the future.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>