Clear serial input buffer after updating display at startup
We issue several $FP commands at startup to display info
on the screen, and each one returns an "$OK" - this causes
the RAPI loop to miscount the successful communications
when it processes these $OK responses for commands it
did not issue, and we see success for i.e. "6 of 3" commands.

Fix this by clearing anything on the input buffer before
we start gathering status.

Signed-off-by: Eric Sandeen <sandeen@sandeen.net>