Don't allow voltage to be set <= 0.0 via MQTT
Voltage is used by divert mode, as well as sent along to the OpenEVSE
module for power calculations.

Setting this to zero could result in a lot of issues. The divert code
would see divide by zero happen, and the energy montoring code on the
OpenEVSE module would result in the recorded energy usage being
incorrect.

While this shouldn't happen reguarly, I found myself in this situation
by sending a null/zero-length message to the relevant topic that the
ESP32 was listening on. Because there is no error checking in the float
parsing code, a null value parsed to zero, which resulted in the voltage
value being set to zero.

An alternative here would be to reset the voltage to `DEFAULT_VOLTAGE`
if a zero or invalid value was received. As written, this will simply do
nothing and ignore the invalid/zero value.