Changed to use ETag/If-None-Match for caching
The date format for the Last-Modified header was wrong and was a bit brute force. Using ETag allows proper caching of the each individual file, even over firmware updates.

Also added the Cache-Control header, to force the browser to check with the WiFi module if the file hase changed before using a cached version.