Added network state management
All network changes are now handled in the main network loop and also added in some state management.

For Wired boards we try to connect only on wired for a period of time before bringing up the WiFi.

In AP mode, if we have client credentials continuosly try to connect the client.

Auto stop the AP after a successful client connection