Fixes to the auto version number script
Updated the auto version number script to use the correct GitHub environment variables to get the version number when running on GitHub Actions.

Added the `_modified` suffix to the version number when the working directory is not clean. This is to ensure that the version number is always different when the working directory is not clean.

Added the `GIT_HASH` define with the short hash of the current commit. This is to make it easier to identify the exact commit that was used to build the binary.

Fixes #584